week-3/DChanHong/requirements.txt
# LangChain 핵심
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# 벡터 스토어
langchain-chroma>=0.1.1
chromadb>=0.5.0

# PDF 처리
pypdf>=4.0.0

# 환경 변수
python-dotenv>=1.0.0
